it works...
Been a hot min since last commit but I've just been too focused on refactoring and making it work.

The readme will explain the program and how to use it.

Added various comments here and there.

But critically the program now works, last commit it didnt save or load anything, now ive brought in json to handle that and added all functionality.