Account Creation and Saving
Account creation and saving is now working properly. the issue with account creation was that in the for loop it didnt add new account to the array and thus couldnt make a new user with the proper Account class to feed into the new User class.

Account saving was a bit difficult to figure out, but the problem was that when i sign in it uses "user" and isnt linked to the users object, to fix it we load the users object from json and search for the right user with the user we sign in with.